+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/backend/engine/NUDState.kt
@@ -1,7 +1,7 @@
 package com.github.warningimhack3r.npmupdatedependencies.backend.engine
 
-import com.github.warningimhack3r.npmupdatedependencies.backend.data.Deprecation
-import com.github.warningimhack3r.npmupdatedependencies.backend.data.ScanResult
+import com.github.warningimhack3r.npmupdatedependencies.backend.data.DeprecationState
+import com.github.warningimhack3r.npmupdatedependencies.backend.data.UpdateState
 import com.github.warningimhack3r.npmupdatedependencies.ui.statusbar.StatusBarHelper
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
@@ -16,15 +16,15 @@ class NUDState {
 
     /**
      * A "cache" of available updates for packages.
-     * Managed by [PackageUpdateChecker.areUpdatesAvailable].
+     * Managed by [PackageUpdateChecker.checkAvailableUpdates] and [com.github.warningimhack3r.npmupdatedependencies.ui.annotation.UpdatesAnnotator.doAnnotate].
      */
-    val availableUpdates = mutableMapOf<String, ScanResult>()
+    val availableUpdates = mutableMapOf<String, UpdateState>()
 
     /**
      * A "cache" of deprecations for packages.
-     * Managed by [com.github.warningimhack3r.npmupdatedependencies.ui.annotation.DeprecationAnnotator.doAnnotate].
+     * Managed by [PackageDeprecationChecker.getDeprecationStatus] and [com.github.warningimhack3r.npmupdatedependencies.ui.annotation.DeprecationAnnotator.doAnnotate].
      */
-    val deprecations = mutableMapOf<String, Deprecation>()
+    val deprecations = mutableMapOf<String, DeprecationState>()
 
     /**
      * A "cache" of registries for packages, mapping package names to registry URLs.